Move state to core directory
