--- core/statediff.go
@@ -1,7 +1,0 @@
-package core
-
-// TODO
-// See the following for an example of a state diff
-// https://alpha-mainnet.starknet.io/feeder_gateway/get_state_update?blockNumber=latest
-type StateDiff struct {
-}
